+++ src/main/java/mekanism/common/item/gear/ItemFlamethrower.java
@@ -60,7 +60,7 @@ public class ItemFlamethrower extends Item implements IItemHUDProvider, IModeIte
     }
 
     @Nonnull
-    public GasStack useGas(ItemStack stack, int amount) {
+    public GasStack useGas(ItemStack stack, long amount) {
         Optional<IGasHandler> capability = MekanismUtils.toOptional(stack.getCapability(Capabilities.GAS_HANDLER_CAPABILITY));
         if (capability.isPresent()) {
             IGasHandler gasHandlerItem = capability.get();
